+++ src/main/java/com/DTISE/ShelfMasterBE/infrastructure/cart/dto/CreateCartItemResponse.java
@@ -1,5 +1,7 @@
 package com.DTISE.ShelfMasterBE.infrastructure.cart.dto;
 
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,11 +14,13 @@ import java.time.OffsetDateTime;
 @AllArgsConstructor
 public class CreateCartItemResponse {
 
-    private BigInteger cartId;
+    private Long cartId;
     private Long productId;
     private String productName;
     private Integer quantity;
     private Boolean isProcessed;
     private OffsetDateTime createdAt;
     private OffsetDateTime updatedAt;
+
+
 }
